requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --all-extras --color requirements-dev.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
astroid==3.3.10
    # via pylint
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
coverage[toml]==7.9.1
    # via pytest-cov
dill==0.4.0
    # via pylint
fastapi==0.115.13
    # via -r /home/streaky/pyformatic-stealth/requirements-demo.in
greenlet==3.2.3
    # via playwright
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r /home/streaky/pyformatic-stealth/requirements-demo.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via pylint
itsdangerous==2.2.0
    # via -r /home/streaky/pyformatic-stealth/requirements-demo.in
jinja2==3.1.6
    # via -r /home/streaky/pyformatic-stealth/requirements.in
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
packaging==25.0
    # via pytest
platformdirs==4.3.8
    # via pylint
playwright==1.52.0
    # via
    #   -r requirements-dev.in
    #   pytest-playwright
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pylint==3.3.7
    # via -r requirements-dev.in
pytest==8.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-base-url
    #   pytest-cov
    #   pytest-playwright
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-cov==6.2.1
    # via -r requirements-dev.in
pytest-playwright==0.7.0
    # via -r requirements-dev.in
python-multipart==0.0.20
    # via -r /home/streaky/pyformatic-stealth/requirements-demo.in
python-slugify==8.0.4
    # via pytest-playwright
pyyaml==6.0.2
    # via -r /home/streaky/pyformatic-stealth/requirements.in
requests==2.32.4
    # via pytest-base-url
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.3
    # via pylint
typing-extensions==4.14.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.34.3
    # via -r /home/streaky/pyformatic-stealth/requirements-demo.in
